(TASKS-39) Remove Pcore modifiable internal state
Prior to this commit, Pcore would hold an instance that maintaned the
loader and implementation registry in a thread local variable. The
design was a less than ideal attempt to mimic the Ruby implementation
where there's no need to provide a Pcore instance.

The thread local is removed in this commit. Instead, a static loader
is created that contains all static types. It is then frozen so that no
furhter types can be defined. This static instance is publicly available
using the static method `Pcore.staticPcore()`.

The method `Pcore.create()` will create a new Pcore instance that is
parented by the static instance. This new instance is modifiable and
its intended use is for creating new types and aliases or to register
implementation for `Object` types.

The removal of the internal state means that the public API of the Pcore
class changed too. All methods aside from the two aformentioned class
methods `staticPcore()` and `create()` have been transformed into
instance methods.