(TASKS-39) Enable hard failure when types are unresolved
This commit adds a boolean argument to the {{Pcore.create}} method that,
when set to `true` will force a hard failure in case a type cannot be
resolved. When `false` (the default), unresolved types are turned into
type references to enable deferred resolution.

The instance returned by {{Pcore.staticPcore}} will have this flag set
to `true` by default.