Add PN parser and make SpecEvaluator use it
This commits adds the following:
- A parser that can parse a string into PN expressions.
- Ability to compare PN expressions for equality.
- Proper PN expression equality comparison in the SpecEvaluator

Since the SpecEvaluator now uses PN expressions when comparing the
expected and actual result, it is now possible to enter such expressions
in a pspec without concern for linebreaks or indentation.

This also means that the the go-parser and pcore-java can use the same
set of pspecs in their tests frameworks.