(PUP-7835) Refactor how TypeMismatchDescriber tracks expected type
The `TypeMismatchDescriber` contained some less ideal logic that made
attempts to revert the effect or using normalized types or types
extracted from an `Optional` or a `TypeAlias`. This commit removes that
code and adds a parameter to get track of the original instead which
results in a simpler and more reliable solution.