(maint) Remove serialization code including jackson-databind and msgpack
This removes all code related to serializing or deserializing pcore
objects, including the JSON and msgpack readers and writers and the
ToDataConverter and FromDataConverter classes.

We do not and have never used this code, as we don't use pcore as a wire
format. jackson-databind is a fairly weighty dependency that also has
frequent security vulnerabilities. Given we don't actually need it,
continuing to maintain that support is not worthwhile.