(maint) Avoid recursively parsing duration formats
Previously, FormatParser.singleton.parseFormat could throw a
ConcurrentModificationException because it was possible for that method
to cause a second nested parseFormat call to happen.

This occurred because of the static DEFAULTS field on DurationFormat,
which contained a list of DurationFormat instances computed by calling
parseFormat on a list of strings. Because it's a static field, it was
computed the first time the class was referenced. Since the FormatParser
class keeps a HashMap<String,DurationFormat>, it was possible for that
first reference (at least in tests) to happen when
FormatParser.singleton.parseFormat was called.

That method uses HashMap.computeIfAbsent to only parse and add the
format if it's not already present. An important requirement for
computeIfAbsent is that the method used to do the computing *must not*
modify the HashMap, or a ConcurrentModificationException can be thrown.
Since the parseFormat method references the "formats" map which
references the DurationFormat, that method could cause an additional
parse to happen during the original parse, throwing that exception.

We now store a list of format strings on DurationFormat (now called
DEFAULT_FORMATS) rather than a list of DurationFormat instances. This
means they will only be computed when *needed*, not when the
DurationFormat class is referenced, avoiding the automatic nesterd calls
to parseFormat. It also brings the defaultParse() method (which parses a
duration with all of the default formats) in line with the parse()
method (which parses a duration with a list of formats), as the latter
method expects a list of strings.

This means that now both the defaultFormat() and defaultParse() methods
are just simple calls to the format() and parse() methods, passing in
defaults.